requirements.txt
streamlit
pandas
langchain
langchain-experimental
langchain-google-genai
python-dotenv
matplotlib
openpyxl
tabulate
